Fixing build error in Nvidia

+++ Source/Math/fpgeneric.h
@@ -19,7 +19,7 @@
 #include "hip/hip_runtime.h"
 #include <hipblas.h>
 #include <hipsparse.h>
-#include <hipDNN.h>
+#include <hipdnn.h>
 #include <hiprand.h>
 #include <hiprand_kernel.h>
 #include <time.h>
@@ -84,7 +84,7 @@ __global__ void GenerateUniformHalf(hiprandState *state, half *result, int n)
     hiprandState localState = *state;
 
     float x;
-    skipahead(id, &localState);
+    skipahead((unsigned long long)id, &localState);
     x = hiprand_uniform(&localState);
 
     result[id] = x;
@@ -99,7 +99,7 @@ __global__ void GenerateNormalHalf(hiprandState *state, half *result, int n, hal
     hiprandState localState = *state;
 
     float x;
-    skipahead(id, &localState);
+    skipahead((unsigned long long)id, &localState);
     x = hiprand_normal(&localState);
 
     result[id] = (float)mean + (float)stddev * x;
